+++ src/color/color.rs
@@ -1,5 +1,4 @@
 use core::fmt;
-use std::collections::HashMap;
 
 use crate::utils;